Version 0.1.1
- Added `equinox.static_field`. This is a bit of a compromise: it's not
quite as flexible as a general "static PyTree node", and because it
somehow privileges behaviour of Modules, it means that the values it
annotates should really _not_ be thought of as being part of a PyTree.
Rather, they're just some extra information that is just kind of there.
On the plus side -- this is the main reason for using it -- it gives an
easy syntax for the common use case.
- As an example of using `static_field`, the various `nn` modules now
store their hyperparameters (width, depth etc.), as a convenience to be
queried later.
- `filter_jit` and `filter_grad` and `filter_value_and_grad` now have
sensible defaults for `filter_spec`. It should probably actually be
relatively unusual to want to change the defaults.
- `GRUCell(bias=)` and `LSTMCell(bias=)` have been deprecated in favour
of `use_bias=`, for consistency with `eqx.nn.Linear`.