Allow ω to take secondary, prefixing tree
jtu.tree_map(fn, tree, *rest) uses tree to determine where to evaluate
f(tree, *rest). In particular, tree must be a prefixing PyTree to
every pytree in  *rest.

This adds support for prefixing trees. in particular:
```
ω(a).call(fn) = jax.tree_util.tree_map(lambda x, y: fn(y), a, a)
```
and
```
ω(a, b).call(fn) = jax.tree_util(lambda x, y: fn(y), b, a)
```

This was included to handle buffers in the equinox internal while loop,
where we may wish to use ω to index into buffers, rather than into the PyTrees
resulting from wrapping subtrees in buffers.

For example, if `a` is a PyTree, and we wrap `a.foo` in a buffer
giving `_Buffer(a)`, then

```
ω(a)[index].ω = ω(buffer_a)[index].ω
```