Added support to structured ω to handle buffers
Allow ω to take two arguments ω(a, struct) where a is the pytree
to which we would to apply ω and struct is a prefixing pytree of a.
This allows us to wrap the leaves of a in buffers and index into
those buffers by passing the appropriate struct.
Specifically
```
ω(a, struct).call(fn) = jax.tree_util.tree_map(lambda x, y:
fn(x[...]), struct, a)

ω(a).call(fn) = ω(a, a).call(fn)
```

Additionally, add support for `ω(a, struct).at[foo].set(bar)` when
a is wrapped in buffers.