Added a MaybeBuffer annotation.
This is for use as a dataclass annotation, as jaxtyping
will verify that the type is correct.

I'm not super happy with this. Buffers are getting increasingly
complicated! I suspect a better resolution will be to move Buffers into
Quax and treat them as array-ish objects, but that's fairly
complicated.